+++ lib/auth/auth_with_roles_test.go
@@ -151,9 +151,13 @@ func TestBotCertificateGenerationCheck(t *testing.T) {
 	t.Parallel()
 	srv := newTestTLSServer(t)
 
+	_, err := CreateRole(context.Background(), srv.Auth(), "example", types.RoleSpecV5{})
+	require.NoError(t, err)
+
 	// Create a new bot.
 	bot, err := srv.Auth().createBot(context.Background(), &proto.CreateBotRequest{
-		Name: "test",
+		Name:  "test",
+		Roles: []string{"example"},
 	})
 	require.NoError(t, err)
 
@@ -204,9 +208,13 @@ func TestBotCertificateGenerationStolen(t *testing.T) {
 	t.Parallel()
 	srv := newTestTLSServer(t)
 
+	_, err := CreateRole(context.Background(), srv.Auth(), "example", types.RoleSpecV5{})
+	require.NoError(t, err)
+
 	// Create a new bot.
 	bot, err := srv.Auth().createBot(context.Background(), &proto.CreateBotRequest{
-		Name: "test",
+		Name:  "test",
+		Roles: []string{"example"},
 	})
 	require.NoError(t, err)
 
@@ -253,6 +261,18 @@ func TestBotCertificateGenerationStolen(t *testing.T) {
 	})
 	require.NoError(t, err)
 	require.NotEmpty(t, locks)
+}
+
+// TestBotNoRoles attempts to create a bot with an empty role list.
+func TestBotNoRoles(t *testing.T) {
+	t.Parallel()
+	srv := newTestTLSServer(t)
+
+	// Create a new bot without roles specified. This should fail.
+	_, err := srv.Auth().createBot(context.Background(), &proto.CreateBotRequest{
+		Name: "test",
+	})
+	require.True(t, trace.IsBadParameter(err))
 }
 
 // TestSSOUserCanReissueCert makes sure that SSO user can reissue certificate
@@ -344,6 +364,7 @@ func TestGenerateUserCertsWithRoleRequest(t *testing.T) {
 		username         string
 		roles            []string
 		roleRequests     []string
+		useRoleRequests  bool
 		expectPrincipals []string
 		expectRoles      []string
 		expectError      func(error) bool
@@ -353,6 +374,7 @@ func TestGenerateUserCertsWithRoleRequest(t *testing.T) {
 			username:         "alice",
 			roles:            []string{emptyRole.GetName(), impersonatorRole.GetName()},
 			roleRequests:     []string{accessFooRole.GetName(), accessBarRole.GetName()},
+			useRoleRequests:  true,
 			expectPrincipals: []string{"foo", "bar"},
 		},
 		{
@@ -360,30 +382,46 @@ func TestGenerateUserCertsWithRoleRequest(t *testing.T) {
 			username:         "bob",
 			roles:            []string{emptyRole.GetName(), impersonatorRole.GetName()},
 			roleRequests:     []string{accessFooRole.GetName()},
+			useRoleRequests:  true,
 			expectPrincipals: []string{"foo"},
 		},
 		{
-			// users not using role requests should keep their own roles
-			desc:         "requesting no roles",
-			username:     "charlie",
-			roles:        []string{emptyRole.GetName()},
-			roleRequests: []string{},
-			expectRoles:  []string{emptyRole.GetName()},
+			// Users not using role requests should keep their own roles
+			desc:            "requesting no roles",
+			username:        "charlie",
+			roles:           []string{emptyRole.GetName()},
+			roleRequests:    []string{},
+			useRoleRequests: false,
+			expectRoles:     []string{emptyRole.GetName()},
 		},
 		{
-			desc:         "requesting a disallowed role",
-			username:     "dave",
-			roles:        []string{emptyRole.GetName()},
-			roleRequests: []string{accessFooRole.GetName()},
+			// An empty role request should fail when role requests are
+			// expected.
+			desc:            "requesting no roles with UseRoleRequests",
+			username:        "charlie",
+			roles:           []string{emptyRole.GetName()},
+			roleRequests:    []string{},
+			useRoleRequests: true,
+			expectError: func(err error) bool {
+				return trace.IsBadParameter(err)
+			},
+		},
+		{
+			desc:            "requesting a disallowed role",
+			username:        "dave",
+			roles:           []string{emptyRole.GetName()},
+			roleRequests:    []string{accessFooRole.GetName()},
+			useRoleRequests: true,
 			expectError: func(err error) bool {
 				return err != nil && trace.IsAccessDenied(err)
 			},
 		},
 		{
-			desc:         "requesting a nonexistent role",
-			username:     "erin",
-			roles:        []string{emptyRole.GetName()},
-			roleRequests: []string{"doesnotexist"},
+			desc:            "requesting a nonexistent role",
+			username:        "erin",
+			roles:           []string{emptyRole.GetName()},
+			roleRequests:    []string{"doesnotexist"},
+			useRoleRequests: true,
 			expectError: func(err error) bool {
 				return err != nil && trace.IsNotFound(err)
 			},
@@ -393,22 +431,25 @@ func TestGenerateUserCertsWithRoleRequest(t *testing.T) {
 			username:         "frank",
 			roles:            []string{emptyRole.GetName(), impersonatorRole.GetName(), denyBarRole.GetName()},
 			roleRequests:     []string{accessFooRole.GetName()},
+			useRoleRequests:  true,
 			expectPrincipals: []string{"foo"},
 		},
 		{
-			desc:         "requesting a denied role",
-			username:     "geoff",
-			roles:        []string{emptyRole.GetName(), impersonatorRole.GetName(), denyBarRole.GetName()},
-			roleRequests: []string{accessBarRole.GetName()},
+			desc:            "requesting a denied role",
+			username:        "geoff",
+			roles:           []string{emptyRole.GetName(), impersonatorRole.GetName(), denyBarRole.GetName()},
+			roleRequests:    []string{accessBarRole.GetName()},
+			useRoleRequests: true,
 			expectError: func(err error) bool {
 				return err != nil && trace.IsAccessDenied(err)
 			},
 		},
 		{
-			desc:         "misusing a role intended for user impersonation",
-			username:     "helen",
-			roles:        []string{emptyRole.GetName(), dummyUserImpersonatorRole.GetName()},
-			roleRequests: []string{dummyUserRole.GetName()},
+			desc:            "misusing a role intended for user impersonation",
+			username:        "helen",
+			roles:           []string{emptyRole.GetName(), dummyUserImpersonatorRole.GetName()},
+			roleRequests:    []string{dummyUserRole.GetName()},
+			useRoleRequests: true,
 			expectError: func(err error) bool {
 				return err != nil && trace.IsAccessDenied(err)
 			},
@@ -434,10 +475,11 @@ func TestGenerateUserCertsWithRoleRequest(t *testing.T) {
 			require.NoError(t, err)
 
 			certs, err := client.GenerateUserCerts(ctx, proto.UserCertsRequest{
-				PublicKey:    pub,
-				Username:     user.GetName(),
-				Expires:      time.Now().Add(time.Hour),
-				RoleRequests: tt.roleRequests,
+				PublicKey:       pub,
+				Username:        user.GetName(),
+				Expires:         time.Now().Add(time.Hour),
+				RoleRequests:    tt.roleRequests,
+				UseRoleRequests: tt.useRoleRequests,
 			})
 			if tt.expectError != nil {
 				require.True(t, tt.expectError(err), "error: %+v: %s", err, trace.DebugReport(err))
@@ -2227,7 +2269,7 @@ func TestListResources_KindKubernetesCluster(t *testing.T) {
 
 	testNames := []string{"a", "b", "c", "d"}
 
-	// Add some kube services.
+	// Add a kube service with 3 clusters.
 	kubeService, err := types.NewServer("bar", types.KindKubeService, types.ServerSpecV2{
 		KubernetesClusters: []*types.KubernetesCluster{{Name: "d"}, {Name: "b"}, {Name: "a"}},
 	})
@@ -2235,7 +2277,8 @@ func TestListResources_KindKubernetesCluster(t *testing.T) {
 	_, err = s.UpsertKubeServiceV2(ctx, kubeService)
 	require.NoError(t, err)
 
-	// Include a duplicate cluster name to test deduplicate.
+	// Add a kube service with 2 clusters.
+	// Includes a duplicate cluster name to test deduplicate.
 	kubeService, err = types.NewServer("foo", types.KindKubeService, types.ServerSpecV2{
 		KubernetesClusters: []*types.KubernetesCluster{{Name: "a"}, {Name: "c"}},
 	})
@@ -2258,7 +2301,8 @@ func TestListResources_KindKubernetesCluster(t *testing.T) {
 		require.NoError(t, err)
 		require.Len(t, res.Resources, len(testNames))
 		require.Empty(t, res.NextKey)
-		require.Empty(t, res.TotalCount)
+		// There is 2 kube services, but 4 unique clusters.
+		require.Equal(t, 4, res.TotalCount)
 
 		clusters, err := types.ResourcesWithLabels(res.Resources).AsKubeClusters()
 		require.NoError(t, err)
@@ -2410,4 +2454,171 @@ func TestDeleteUserAppSessions(t *testing.T) {
 	sessions, err = srv.Auth().GetAppSessions(ctx)
 	require.NoError(t, err)
 	require.Len(t, sessions, 0)
+}
+
+func TestListResources_SortAndDeduplicate(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
+	srv := newTestTLSServer(t)
+
+	// Create user, role, and client.
+	username := "user"
+	user, role, err := CreateUserAndRole(srv.Auth(), username, nil)
+	require.NoError(t, err)
+	identity := TestUser(user.GetName())
+	clt, err := srv.NewClient(identity)
+	require.NoError(t, err)
+
+	// Permit user to get all resources.
+	role.SetWindowsDesktopLabels(types.Allow, types.Labels{types.Wildcard: {types.Wildcard}})
+	require.NoError(t, srv.Auth().UpsertRole(ctx, role))
+
+	// Define some resource names for testing.
+	names := []string{"d", "b", "d", "a", "a", "b"}
+	uniqueNames := []string{"a", "b", "d"}
+
+	tests := []struct {
+		name            string
+		kind            string
+		insertResources func()
+		wantNames       []string
+	}{
+		{
+			name: "KindDatabaseServer",
+			kind: types.KindDatabaseServer,
+			insertResources: func() {
+				for i := 0; i < len(names); i++ {
+					db, err := types.NewDatabaseServerV3(types.Metadata{
+						Name: fmt.Sprintf("name-%v", i),
+					}, types.DatabaseServerSpecV3{
+						HostID:   "_",
+						Hostname: "_",
+						Database: &types.DatabaseV3{
+							Metadata: types.Metadata{
+								Name: names[i],
+							},
+							Spec: types.DatabaseSpecV3{
+								Protocol: "_",
+								URI:      "_",
+							},
+						},
+					})
+					require.NoError(t, err)
+					_, err = srv.Auth().UpsertDatabaseServer(ctx, db)
+					require.NoError(t, err)
+				}
+			},
+		},
+		{
+			name: "KindAppServer",
+			kind: types.KindAppServer,
+			insertResources: func() {
+				for i := 0; i < len(names); i++ {
+					server, err := types.NewAppServerV3(types.Metadata{
+						Name: fmt.Sprintf("name-%v", i),
+					}, types.AppServerSpecV3{
+						HostID: "_",
+						App:    &types.AppV3{Metadata: types.Metadata{Name: names[i]}, Spec: types.AppSpecV3{URI: "_"}},
+					})
+					require.NoError(t, err)
+					_, err = srv.Auth().UpsertApplicationServer(ctx, server)
+					require.NoError(t, err)
+				}
+			},
+		},
+		{
+			name: "KindWindowsDesktop",
+			kind: types.KindWindowsDesktop,
+			insertResources: func() {
+				for i := 0; i < len(names); i++ {
+					desktop, err := types.NewWindowsDesktopV3(names[i], nil, types.WindowsDesktopSpecV3{
+						Addr:   "_",
+						HostID: fmt.Sprintf("name-%v", i),
+					})
+					require.NoError(t, err)
+					require.NoError(t, srv.Auth().UpsertWindowsDesktop(ctx, desktop))
+				}
+			},
+		},
+		{
+			name: "KindKubernetesCluster",
+			kind: types.KindKubernetesCluster,
+			insertResources: func() {
+				for i := 0; i < len(names); i++ {
+					server, err := types.NewServer(fmt.Sprintf("name-%v", i), types.KindKubeService, types.ServerSpecV2{
+						KubernetesClusters: []*types.KubernetesCluster{
+							// Test dedup inside this list as well as from each service.
+							{Name: names[i]},
+							{Name: names[i]},
+						},
+					})
+					require.NoError(t, err)
+					_, err = srv.Auth().UpsertKubeServiceV2(ctx, server)
+					require.NoError(t, err)
+				}
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			tc.insertResources()
+
+			// Fetch all resources
+			fetchedResources := make([]types.ResourceWithLabels, 0, len(uniqueNames))
+			resp, err := clt.ListResources(ctx, proto.ListResourcesRequest{
+				ResourceType:   tc.kind,
+				NeedTotalCount: true,
+				Limit:          2,
+				SortBy:         types.SortBy{Field: types.ResourceMetadataName, IsDesc: true},
+			})
+			require.NoError(t, err)
+			require.Len(t, resp.Resources, 2)
+			require.Equal(t, len(uniqueNames), resp.TotalCount)
+			fetchedResources = append(fetchedResources, resp.Resources...)
+
+			resp, err = clt.ListResources(ctx, proto.ListResourcesRequest{
+				ResourceType:   tc.kind,
+				NeedTotalCount: true,
+				StartKey:       resp.NextKey,
+				Limit:          2,
+				SortBy:         types.SortBy{Field: types.ResourceMetadataName, IsDesc: true},
+			})
+			require.NoError(t, err)
+			require.Len(t, resp.Resources, 1)
+			require.Equal(t, len(uniqueNames), resp.TotalCount)
+			fetchedResources = append(fetchedResources, resp.Resources...)
+
+			r := types.ResourcesWithLabels(fetchedResources)
+			var extractedErr error
+			var extractedNames []string
+
+			switch tc.kind {
+			case types.KindDatabaseServer:
+				s, err := r.AsDatabaseServers()
+				require.NoError(t, err)
+				extractedNames, extractedErr = types.DatabaseServers(s).GetFieldVals(types.ResourceMetadataName)
+
+			case types.KindAppServer:
+				s, err := r.AsAppServers()
+				require.NoError(t, err)
+				extractedNames, extractedErr = types.AppServers(s).GetFieldVals(types.ResourceMetadataName)
+
+			case types.KindWindowsDesktop:
+				s, err := r.AsWindowsDesktops()
+				require.NoError(t, err)
+				extractedNames, extractedErr = types.WindowsDesktops(s).GetFieldVals(types.ResourceMetadataName)
+
+			default:
+				s, err := r.AsKubeClusters()
+				require.NoError(t, err)
+				require.Len(t, s, 3)
+				extractedNames, extractedErr = types.KubeClusters(s).GetFieldVals(types.ResourceMetadataName)
+			}
+
+			require.NoError(t, extractedErr)
+			require.ElementsMatch(t, uniqueNames, extractedNames)
+			require.IsDecreasing(t, extractedNames)
+		})
+	}
 }