+++ lib/kube/proxy/remotecommand.go
@@ -22,6 +22,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"strings"
 	"time"
 
 	apievents "github.com/gravitational/teleport/api/types/events"
@@ -203,6 +204,10 @@ func (s *remoteCommandProxy) sendStatus(err error) error {
 			Status: metav1.StatusSuccess,
 		}})
 	}
+	if statusErr, ok := err.(*apierrors.StatusError); ok {
+		return s.writeStatus(statusErr)
+	}
+
 	if exitErr, ok := err.(utilexec.ExitError); ok && exitErr.Exited() {
 		rc := exitErr.ExitStatus()
 		return s.writeStatus(&apierrors.StatusError{ErrStatus: metav1.Status{
@@ -219,6 +224,22 @@ func (s *remoteCommandProxy) sendStatus(err error) error {
 			Message: fmt.Sprintf("command terminated with non-zero exit code: %v", exitErr),
 		}})
 	}
+	// kubernetes client-go errorDecoderV4 parses the metav1.Status and returns the `fmt.Errorf(status.Message)` for every case except
+	// errors with reason =  NonZeroExitCodeReason for which it returns an exec.CodeExitError.
+	// This means when forwarding an exec request to a remote cluster using the `Forwarder.remoteExec` function we only have access
+	// to the status.Message. This happens because the error is sent after the connection was upgraded to a bidirectional stream.
+	// This hack is here to recreate the forbidden message and return it back to the user terminal
+	if strings.Contains(err.Error(), "is forbidden:") {
+		return s.writeStatus(&apierrors.StatusError{
+			ErrStatus: metav1.Status{
+				Status:  metav1.StatusFailure,
+				Code:    http.StatusForbidden,
+				Reason:  metav1.StatusReasonForbidden,
+				Message: err.Error(),
+			},
+		})
+	}
+
 	err = trace.BadParameter("error executing command in container: %v", err)
 	return s.writeStatus(apierrors.NewInternalError(err))
 }