+++ lib/tbot/renew.go
@@ -135,13 +135,13 @@ func describeTLSIdentity(ident *identity.Identity) (string, error) {
 // identityConfigurator is a function that alters a cert request
 type identityConfigurator = func(req *proto.UserCertsRequest)
 
-// generateIdentity uses an identity to retrieve another possibly impersonated
-// identity. The `configurator` function, if not nil, can be used to add
-// additional requests to the certificate request, for example to add
-// `RouteToDatabase` and similar fields, however in that case it must be
-// called with an impersonated identity that already has the relevant
-// permissions, much like `tsh (app|db|kube) login` is already used to generate
-// an additional set of certs.
+// generateIdentity uses an identity to retrieve an impersonated identity.
+// The `configurator` function, if not nil, can be used to add additional
+// requests to the certificate request, for example to add `RouteToDatabase`
+// and similar fields, however in that case it must be called with an
+// impersonated identity that already has the relevant permissions, much like
+// `tsh (app|db|kube) login` is already used to generate an additional set of
+// certs.
 func (b *Bot) generateIdentity(
 	ctx context.Context,
 	currentIdentity *identity.Identity,
@@ -174,6 +174,11 @@ func (b *Bot) generateIdentity(
 		Expires:        expires,
 		RoleRequests:   roleRequests,
 		RouteToCluster: currentIdentity.ClusterName,
+
+		// Make sure to specify this is an impersonated cert request. If unset,
+		// auth cannot differentiate renewable vs impersonated requests when
+		// len(roleRequests) == 0.
+		UseRoleRequests: true,
 	}
 
 	if configurator != nil {
@@ -304,6 +309,8 @@ func (b *Bot) generateImpersonatedIdentity(
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
+
+		defer impClient.Close()
 
 		route, err := b.getRouteToDatabase(ctx, impClient, destCfg.Database)
 		if err != nil {
@@ -463,9 +470,9 @@ func (b *Bot) renew(
 		return trace.Wrap(err, "unable to communicate with auth server")
 	}
 
-	b.log.Debug("Auth client now using renewed credentials.")
 	b.setClient(newClient)
 	b.setIdent(newIdentity)
+	b.log.Debug("Auth client now using renewed credentials.")
 
 	// Now that we're sure the new creds work, persist them.
 	if err := identity.SaveIdentity(newIdentity, botDestination, identity.BotKinds()...); err != nil {